requirements.txt
flask==3.1.0
anthropic==0.49.0
python-dotenv==1.0.1
pytest==8.3.5
pytest-flask==1.3.0
